requirements.txt
# Django and REST framework
Django>=5.0.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.5

# Environment variables
python-dotenv>=1.0.0

# OpenAI
openai>=1.0.0

# Celery and Redis
celery>=5.3.0
redis>=4.5.0

# HTTP requests
requests>=2.28.0

# CORS headers
django-cors-headers>=4.0.0

# UI Enhancement
django-tailwind>=3.6.0
django-htmx>=1.16.0
django-browser-reload>=1.11.0

# Debugging
django-debug-toolbar>=4.0.0

# Testing
pytest>=7.0.0
pytest-django>=4.5.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
Sphinx>=6.0.0
